Move core schemas to core sub package
