+++ jolpica_api/f1_api/serializers/driver.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from rest_framework import serializers
 
 from jolpica.formula_one import models as f1
-from jolpica_schemas.f1_api.alpha.driver import DriverSummary
+from jolpica_schemas.f1_api.alpha.core.driver import DriverSummary
 
 from .base_serializer import BaseAPISerializer
