+++ jolpica_api/f1_api/serializers/lap.py
@@ -4,7 +4,7 @@ from rest_framework import serializers
 
 from jolpica.formula_one import models as f1
 from jolpica.formula_one.utils import format_timedelta, timedelta_to_iso8601
-from jolpica_schemas.f1_api.alpha.lap import LapPitStop, LapSessionEntry, LapSummary
+from jolpica_schemas.f1_api.alpha.core.lap import LapPitStop, LapSessionEntry, LapSummary
 
 from .base_serializer import BaseAPISerializer
