+++ jolpica_api/f1_api/serializers/season.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from jolpica.formula_one import models as f1
-from jolpica_schemas.f1_api.alpha.season import SeasonSummary
+from jolpica_schemas.f1_api.alpha.core.season import SeasonSummary
 
 from .base_serializer import BaseAPISerializer
