+++ jolpica_api/f1_api/views/seasons.py
@@ -1,7 +1,7 @@
 from drf_spectacular.utils import extend_schema, extend_schema_view
 
 from jolpica.formula_one import models as f1
-from jolpica_schemas.f1_api.alpha.season import RetrievedSeasonDetail, SeasonQueryParams, SeasonSummary
+from jolpica_schemas.f1_api.alpha.core.season import RetrievedSeasonDetail, SeasonQueryParams, SeasonSummary
 
 from ..serializers.season import SeasonSerializer
 from ..utils import pydantic_to_open_api_parameters
