+++ jolpica_api/f1_api/views/session_entries.py
@@ -1,7 +1,7 @@
 from drf_spectacular.utils import extend_schema, extend_schema_view
 
 from jolpica.formula_one import models as f1
-from jolpica_schemas.f1_api.alpha.session_entry import (
+from jolpica_schemas.f1_api.alpha.core.session_entry import (
     RetrievedSessionEntryDetail,
     SessionEntryQueryParams,
     SessionEntrySummary,
