+++ libraries/jolpica-schemas/src/jolpica_schemas/f1_api/alpha/core/session.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, ConfigDict, Field
 
-from .metadata import DetailResponse, PaginatedResponse
-from .shared import BasicRound, Session
+from ..metadata import DetailResponse, PaginatedResponse
+from ..shared import BasicRound, Session
 
 
 class SessionRound(BasicRound):
