+++ libraries/jolpica-schemas/src/jolpica_schemas/f1_api/alpha/core/session_entry.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, ConfigDict, Field
 
-from .metadata import DetailResponse, PaginatedResponse
-from .shared import BasicDriver, BasicRound, BasicSession, BasicTeam, SessionEntry
+from ..metadata import DetailResponse, PaginatedResponse
+from ..shared import BasicDriver, BasicRound, BasicSession, BasicTeam, SessionEntry
 
 
 class SessionEntrySession(BasicSession):
